# Conflicts:
#	RockWeb/Blocks/BulkImport/CsvImport.ascx.cs

+++ RockWeb/Blocks/BulkImport/CsvImport.ascx.cs
@@ -131,8 +131,12 @@ namespace RockWeb.Blocks.BulkImport
             if ( !Page.IsPostBack )
             {
                 // delete all the csv files in the root directory on start up to ensure that no residual files are present before the upload
-                Directory.EnumerateFiles( Request.MapPath( fupCSVFile.RootFolder ), "*.csv" ).ToList()
-                    .ForEach( f => File.Delete( f ) );
+                string directoryPath = Request.MapPath( fupCSVFile.RootFolder );
+                if ( Directory.Exists( directoryPath ) )
+                {
+                    Directory.EnumerateFiles( directoryPath, "*.csv" ).ToList()
+                        .ForEach( f => File.Delete( f ) );
+                }
             }
             RockPage.AddScriptLink( "~/Scripts/jquery.signalR-2.2.0.min.js", false );
         }
@@ -209,6 +213,13 @@ namespace RockWeb.Blocks.BulkImport
 
         protected void btnStart_Click( object sender, EventArgs e )
         {
+            if ( hfCSVFileName.Value.IsNullOrWhiteSpace() )
+            {
+                nbDuplicateHeadersInFile.Text = $"A CSV file must be selected first.";
+                nbDuplicateHeadersInFile.Visible = true;
+                return;
+            }
+
             string csvFileName = this.Request.MapPath( hfCSVFileName.Value );
 
             this.propertiesDropDownList = CreateListItemsDropDown();
@@ -228,7 +239,7 @@ namespace RockWeb.Blocks.BulkImport
                 bool headerContainsDuplicate = !string.IsNullOrEmpty( duplicateHeadersList );
                 if ( headerContainsDuplicate )
                 {
-                    nbDuplicateHeadersInFile.Text = $"The File has the duplicated headers: {duplicateHeadersList}. Please fix it and upload again.";
+                    nbDuplicateHeadersInFile.Text = $"The file has duplicated headers: {duplicateHeadersList}. Please fix it and upload again.";
                     nbDuplicateHeadersInFile.Visible = true;
                     return;
                 }
